refactoring :: UI__+(func) 세분화
