+++ Projects/Modules/DesignSystem/Sources/Extensions/UIButton/UIButton+setUnderline.swift
@@ -1,0 +1,14 @@
+import UIKit
+
+public extension UIButton {
+    func setUnderline() {
+        guard let title = title(for: .normal) else { return }
+        let attributedString = NSMutableAttributedString(string: title)
+        attributedString.addAttribute(
+            .underlineStyle,
+            value: NSUnderlineStyle.single.rawValue,
+            range: NSRange(location: 0, length: title.count)
+        )
+        setAttributedTitle(attributedString, for: .normal)
+    }
+}
