+++ Projects/Modules/DesignSystem/Sources/Extensions/UIImage/UIImage+resize.swift
@@ -2,24 +2,19 @@ import UIKit
 
 public extension UIImage {
     func resize(_ targetSize: CGSize) -> UIImage {
-        // Determine the scale factor that preserves aspect ratio
         let widthRatio = targetSize.width / size.width
         let heightRatio = targetSize.height / size.height
 
         let scaleFactor = min(widthRatio, heightRatio)
 
-        // Compute the new image size that preserves aspect ratio
         let scaledImageSize = CGSize(
             width: size.width * scaleFactor,
             height: size.height * scaleFactor
         )
 
-        // Draw and return the resized UIImage
-        let renderer = UIGraphicsImageRenderer(
+        let scaledImage = UIGraphicsImageRenderer(
             size: scaledImageSize
-        )
-
-        let scaledImage = renderer.image { _ in
+        ).image { _ in
             self.draw(in: CGRect(
                 origin: .zero,
                 size: scaledImageSize
